Minor fixes and separate dashboard for volunteer and admin
