+++ frontend/src/pages/dashboard/VolunteerDashboard.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import "../../styles/pagesStyles/dashboard/Dashboard.css";
+
+function VolunteerDashboard(){
+  return (
+    <>
+      <div className="navgap"></div>
+      <h1>Volunteer Dashboard</h1>
+    </>
+  )
+}
+
+export default VolunteerDashboard;
